Changed Data types

+++ ECommerceAppSwiftUI/ECommerceAppSwiftUI/Views/Bag/Success/SuccessView.swift
@@ -39,7 +39,7 @@ struct SuccessView: View {
                 return
             }
             
-            let total = AppConstants.cartList.reduce(0, { $0 + ($1.price - ($1.price * $1.discount)/100)})
+            let total = Float(AppConstants.cartList.reduce(0, { $0 + ($1.price - ($1.price * $1.discount)/100)}))
             
             let firstProduct = AppConstants.cartList[0]
